validate token & private routes

+++ routes/sites.js
@@ -2,8 +2,11 @@ const express = require('express')
 const router = express.Router()
 const Site = require('../models/site')
 const googleDrive = require('../controllers/googleDrive')
+const verifyToken = require('../middlewares/verifyToken')
 
+// PUBLIC routes
 // Get sites
+// TODO: add filter by owner
 router.get('/', async (req, res) => {
   try {
     const sites = await Site.find()
@@ -24,13 +27,13 @@ router.get('/:siteId', async (req, res) => {
   }
 })
 
+// PRIVATE Routes
 // Create a site
-// TODO: validate user logged in
-router.post('/', async (req, res) => {
+router.post('/', verifyToken.private, async (req, res) => {
   const site = new Site({
     title: req.body.title,
     description: req.body.description,
-    owner: req.body.owner, // TODO get current logued in user
+    owner: req.user._id,
     googleFolderId: req.body.googleFolderId
   })
 
@@ -43,14 +46,19 @@ router.post('/', async (req, res) => {
 })
 
 // Update a site
-// TODO: user is owner of the site
-router.patch('/:siteId', async (req, res) => {
+router.patch('/:siteId', verifyToken.private, async (req, res) => {
+  const site = await Site.findOne({ _id: req.params.siteId })
+
+  if (site && req.user._id != site.owner)
+    return res.status(401).send('You are not allowed to modify the site')
+
   try {
     const updatedSite = await Site.updateOne(
       { _id: req.params.siteId },
       { $set: {
         title: req.body.title,
-        description: req.body.description
+        description: req.body.description,
+        googleFolderId: req.body.googleFolderId
       }}
     )
     res.json(updatedSite)
@@ -61,7 +69,12 @@ router.patch('/:siteId', async (req, res) => {
 
 // Delete a site
 // TODO: user is owner of the site
-router.delete('/:siteId', async (req, res) => {
+router.delete('/:siteId', verifyToken.private, async (req, res) => {
+  const site = await Site.findOne({ _id: req.params.siteId })
+
+  if (site && req.user._id != site.owner)
+    return res.status(401).send('You are not allowed to delete the site')
+
   try {
     const removedSite = await Site.deleteOne({ _id: req.params.siteId })
     res.json(removedSite)
